Added more changes to complete the Release build.

+++ neo/d3xp/Entity.cpp
@@ -3614,7 +3614,7 @@ bool idEntity::HandleGuiCommands( idEntity *entityGui, const char *cmds ) {
 	if ( entityGui && cmds && *cmds ) {
 		idLexer src;
 		idToken token, token2, token3, token4;
-		src.LoadMemory( cmds, strlen( cmds ), "guiCommands" );
+		src.LoadMemory( cmds, static_cast<int>(strlen( cmds )), "guiCommands" );
 		while( 1 ) {
 
 			if ( !src.ReadToken( &token ) ) {
@@ -3944,7 +3944,7 @@ idCurve_Spline<idVec3> *idEntity::GetSpline() const {
 		return NULL;
 	}
 
-	idStr str = kv->GetKey().Right( kv->GetKey().Length() - strlen( curveTag ) );
+	idStr str = kv->GetKey().Right( kv->GetKey().Length() - static_cast<int>(strlen( curveTag )) );
 	if ( str.Icmp( "CatmullRomSpline" ) == 0 ) {
 		spline = new (TAG_ENTITY) idCurve_CatmullRomSpline<idVec3>();
 	} else if ( str.Icmp( "nubs" ) == 0 ) {
