+++ neo/d3xp/gamesys/Event.h
@@ -60,7 +60,7 @@ private:
 	int							returnType;
 	int							numargs;
 	size_t						argsize;
-	int							argOffset[ D_EVENT_MAXARGS ];
+	size_t						argOffset[ D_EVENT_MAXARGS ];
 	int							eventnum;
 	const idEventDef *			next;
 
@@ -77,7 +77,7 @@ public:
 	int							GetEventNum() const;
 	int							GetNumArgs() const;
 	size_t						GetArgSize() const;
-	int							GetArgOffset( int arg ) const;
+	size_t						GetArgOffset( int arg ) const;
 
 	static int					NumEventCommands();
 	static const idEventDef		*GetEventCommand( int eventnum );
@@ -195,7 +195,7 @@ ID_INLINE size_t idEventDef::GetArgSize() const {
 idEventDef::GetArgOffset
 ================
 */
-ID_INLINE int idEventDef::GetArgOffset( int arg ) const {
+ID_INLINE size_t idEventDef::GetArgOffset( int arg ) const {
 	assert( ( arg >= 0 ) && ( arg < D_EVENT_MAXARGS ) );
 	return argOffset[ arg ];
 }
