+++ neo/framework/Common.cpp
@@ -653,7 +653,7 @@ void idCommonLocal::FilterLangList( idStrList* list, idStr lang ) {
 	idStr temp;
 	for( int i = 0; i < list->Num(); i++ ) {
 		temp = (*list)[i];
-		temp = temp.Right(temp.Length()-strlen("strings/"));
+		temp = temp.Right(temp.Length()- static_cast<int>(strlen("strings/")));
 		temp = temp.Left(lang.Length());
 		if(idStr::Icmp(temp, lang) != 0) {
 			list->RemoveIndex(i);
@@ -1013,7 +1013,7 @@ void idCommonLocal::Init( int argc, const char * const * argv, const char *cmdli
 		stringsFile.SetNameAndType( SAVEGAME_STRINGS_FILENAME, SAVEGAMEFILE_BINARY );
 		stringsFile.PreAllocate( MAX_SAVEGAME_STRING_TABLE_SIZE );
 
-		fileSystem->BeginLevelLoad( "_startup", saveFile.GetDataPtr(), saveFile.GetAllocated() );
+		fileSystem->BeginLevelLoad( "_startup", saveFile.GetDataPtr(), static_cast<int>(saveFile.GetAllocated()) );
 
 		// initialize the declaration manager
 		declManager->Init();
